basic_async_syntax

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -3,7 +3,7 @@
 import asyncio, random
 
 
-async def wait_random(max_delay: int = 8) -> int:
+async def wait_random(max_delay: int = 8) -> float:
     """Waiting for a random num of seconds"""
     w_time = random.random() * max_delay
     await asyncio.sleep(w_time)
